[vinh] refactoring folders
