+++ src/main/java/com/abc/accounts/SavingAccount.java
@@ -1,6 +1,6 @@
 package com.abc.accounts;
 
-import com.abc.accounts.interestRateCalculator.InterestRateCalculator;
+import com.abc.interestRateCalculators.InterestRateCalculator;
 
 public class SavingAccount extends AbstractAccount {
