--- src/main/java/com/abc/accounts/interestRateCalculator/InterestRateCalculator.java
@@ -1,11 +1,0 @@
-package com.abc.accounts.interestRateCalculator;
-
-import com.abc.Transaction;
-
-import java.util.List;
-
-public interface InterestRateCalculator<T> {
-    double calculateInterestRate(T input);
-
-
-}
