+++ src/main/java/com/abc/interestRateCalculators/InterestRateCalculator.java
@@ -1,0 +1,7 @@
+package com.abc.interestRateCalculators;
+
+public interface InterestRateCalculator<T> {
+    double calculateInterestRate(T input);
+
+
+}
