+++ src/test/java/com/abc/BankTest.java
@@ -1,8 +1,8 @@
 package com.abc;
 
 import com.abc.accounts.*;
-import com.abc.accounts.interestRateCalculator.StubAmountInterestRateCalculator;
-import com.abc.accounts.interestRateCalculator.StubInterestRateCalculator;
+import com.abc.interestRateCalculators.StubAmountInterestRateCalculator;
+import com.abc.interestRateCalculators.StubInterestRateCalculator;
 import org.hamcrest.CoreMatchers;
 import org.junit.Assert;
 import org.junit.Test;
