+++ src/test/java/com/abc/accounts/AbstractAccountTest.java
@@ -1,6 +1,6 @@
 package com.abc.accounts;
 
-import com.abc.accounts.interestRateCalculator.StubInterestRateCalculator;
+import com.abc.interestRateCalculators.StubInterestRateCalculator;
 import org.hamcrest.CoreMatchers;
 import org.junit.Assert;
 import org.junit.Test;
