+++ src/test/java/com/abc/interestRateCalculators/CheckingInterestRateCalculatorTest.java
@@ -1,4 +1,4 @@
-package com.abc.accounts.interestRateCalculator;
+package com.abc.interestRateCalculators;
 
 import com.abc.Transaction;
 import org.hamcrest.CoreMatchers;
